création d'une deuxième page
