+++ includes/variables.php
@@ -1,0 +1,24 @@
+<?php
+    $myProducts = [
+        [
+            'name' => "Bonnet en laine",
+            'price' => 10,
+            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.",
+            'src' => "https://maison-joseph.fr/1198-thickbox_default/bonnet-laine-et-fourrure.jpg"
+        ], [
+            'name' => "Bonnet en bio",
+            'price' => 14,
+            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.",
+            'src' => "https://www.altermundi.com/395470-thickbox_default/bonnet-en-laine-bio-cotelee-leaf.jpg"
+        ], [
+            'name' => "Bonnet en laine et cachemire",
+            'price' => 20,
+            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.",
+            'src' => "https://m.media-amazon.com/images/I/71oJSr9TGOL._AC_UY445_.jpg"
+        ], [
+            'name' => "Bonnet arc-en-ciel",
+            'price' => 12,
+            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.",
+            'src' => "https://rainbow-blog.com/wp-content/uploads/2018/03/bonnet-tricot-arc-en-ciel-fleur-300x300.jpg"
+        ]
+    ];
