--- variables.php
@@ -1,20 +1,0 @@
-<?php
-    $myProducts = [
-        [
-            'name' => "Bonnet en laine",
-            'price' => 10,
-            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida."
-        ], [
-            'name' => "Bonnet en bio",
-            'price' => 14,
-            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida."
-        ], [
-            'name' => "Bonnet en laine et cachemire",
-            'price' => 20,
-            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida."
-        ], [
-            'name' => "Bonnet arc-en-ciel",
-            'price' => 12,
-            'desc' => "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida."
-        ]
-    ];
